reconfiguracao do app para buscar os dados na API

+++ app/src/main/java/br/com/fuzusnoary/passrepository/listeners/ClickList.java
@@ -1,6 +1,6 @@
 package br.com.fuzusnoary.passrepository.listeners;
 
 public interface ClickList {
-    void onClick(int id);
-    void onDelete(int id);
+    void onClick(long id);
+    void onDelete(long id);
 }
